added  fn  search-genres-and-year
