+++ src/js/renderPopularFilm.js
@@ -1,6 +1,8 @@
 import API from './apiService';
 import videoCardTpl from '../partials/hbs/video-card.hbs'
 import refs from './common/refs';
+import { onSearchYear } from './searchGenresAndYear.js'
+import {onSearchGenresList} from './searchGenresAndYear.js'
 
 const {galleryListRefs, btnHomeRefs} = refs.refs
 
@@ -13,10 +15,8 @@ export default function onSearchPopularFilms(e, page) {
     }
 
     API.fetchGenres().then(data => {
-        // console.log(data.genres);
         return data.genres
     }).then(onSaveGenres)
-
     
     API.fetchPopularFilms(page)
         .then(onSearchYear)
@@ -43,55 +43,3 @@ onSearchPopularFilms()
 function onSaveGenres(genres) {
   localStorage.setItem('genres', JSON.stringify(genres));
 }
-
-
-function onSearchGenresList(data) {
-    const genres = JSON.parse(localStorage.getItem('genres'));
-    data.results.map(item => {
-        let genresArr = [];
-        genres.find(genr => {
-      if (item.genre_ids.includes(genr.id)) {
-        genresArr.push(genr.name)
-      }
-        });
-        
-        if (genresArr.length <= 3) {
-            genresArr = genresArr.join(', ').split(',')
-        }
-        if (genresArr.length >= 3) {
-            genresArr = genresArr.slice(0, 2)
-            genresArr.push(' Other')
-        }
-        
-        
-        item.genre_ids = genresArr
-
-    });
-    // console.log(genresArr);
-    
-
-  return data;
-}
-
-
-
-
-
-
-
-//----------------
-function onSearchYear(data) {
-  data.results.map(elem => {
-    if (elem.release_date) {
-      elem.release_date = elem.release_date.split('-')[0];
-        // console.log(elem.release_date);
-    } else {
-        elem.release_date = 'No date'
-    }
-  })
-
-  return data;
-}
-//------------------------
-
-
